refactor(orchestration): extract HumanQuestionService
Eight methods, 681 body lines, moved verbatim. 2325 -> 1704 lines. The largest
leaf service, and the one the spec explicitly declines to split further: its
methods share the coordinatorAnswerQuestion/workerRaiseQuestion state machine,
and cutting it along a coordinator/worker seam would only create round trips.

The plan named three collaborators and missed TaskCancellationService, which
coordinatorRetractAnswer needs -- that is why Task 14 made startWorkerCancellation
public. Its port list also named deliverCoordinatorMessage, owned by
QuestionFlowCore, and logger, which nothing uses. Six ports, four collaborators.

Two calls keep their position outside the enclosing mutate: the reconcile at the
tail of coordinatorReviewContestedResult, and the bare unawaited
startWorkerCancellation in coordinatorRetractAnswer. Both already pointed
straight at their collaborator rather than through a facade delegation, which is
what keeps the detached chain's hop depth -- and the fixtures that pin its
interleaving -- unchanged.

Closes a real coverage gap on the way past. coordinatorReviewContestedResult
re-arms a pending notice under a four-way compound guard, and the
`noticeSentAt !== undefined` sub-case -- a notice already sent, so noticePending
must not be re-armed -- was pinned by nothing in the suite. Deleting that
conjunct leaves all 185 oracle tests and all 24 golden fixtures green. Two tests
now cover both sides; the mutation turns exactly one of them red.
